requirements.txt
langchain>=0.3.0
langchain-core>=0.3.0
langchain-google-genai>=2.0.0
langchain-chroma>=0.1.0
chromadb>=0.5.0
pydantic>=2.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
pytest>=8.0.0
